feat(images): implemented QA design for the page header
https://warthogs.atlassian.net/browse/MAASENG-4359
