--- src/app/images/components/SMImagesTable/ImagesTableHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ImagesTableHeader";
