+++ src/app/images/components/SMImagesTable/SMImagesTable.tsx
@@ -1,10 +1,8 @@
-import React, { useState } from "react";
+import type { Dispatch, SetStateAction } from "react";
 
 import { TableCaption } from "@canonical/maas-react-components";
 import type { RowSelectionState } from "@tanstack/react-table";
 import { useSelector } from "react-redux";
-
-import ImagesTableHeader from "./ImagesTableHeader";
 
 import GenericTable from "@/app/base/components/GenericTable";
 import { useSidePanel } from "@/app/base/side-panel-context";
@@ -20,6 +18,11 @@ import { splitResourceName } from "@/app/store/bootresource/utils";
 import configSelectors from "@/app/store/config/selectors";
 
 import "./_index.scss";
+
+type SMImagesTableProps = {
+  selectedRows: RowSelectionState;
+  setSelectedRows: Dispatch<SetStateAction<RowSelectionState>>;
+};
 
 const getImages = (resources: BootResource[]): Image[] => {
   return resources.map((resource) => {
@@ -40,8 +43,10 @@ const getImages = (resources: BootResource[]): Image[] => {
   });
 };
 
-export const SMImagesTable: React.FC = () => {
-  const ubuntu = useSelector(bootResourceSelectors.ubuntu);
+const SMImagesTable = ({
+  selectedRows,
+  setSelectedRows,
+}: SMImagesTableProps) => {
   const resources = useSelector(bootResourceSelectors.resources);
   const images = getImages(resources);
 
@@ -50,8 +55,6 @@ export const SMImagesTable: React.FC = () => {
   const commissioningRelease = useSelector(
     configSelectors.commissioningDistroSeries
   );
-
-  const [selectedRows, setSelectedRows] = useState<RowSelectionState>({});
 
   const columns = useImageTableColumns({
     commissioningRelease,
@@ -63,7 +66,7 @@ export const SMImagesTable: React.FC = () => {
         setSidePanelContent({
           view: ImageSidePanelViews.DELETE_MULTIPLE_IMAGES,
           extras: {
-            rowSelection: { ...selectedRows, [row.original.id]: true },
+            rowSelection: { ...selectedRows, [row.id]: true },
             setRowSelection: setSelectedRows,
           },
         });
@@ -72,39 +75,31 @@ export const SMImagesTable: React.FC = () => {
   });
 
   return (
-    <>
-      {!!ubuntu && (
-        <ImagesTableHeader
-          selectedRows={selectedRows}
-          setSelectedRows={setSelectedRows}
-        />
-      )}
-      <GenericTable
-        canSelect
-        columns={columns}
-        data={images}
-        filterCells={filterCells}
-        filterHeaders={filterHeaders}
-        groupBy={["name"]}
-        noData={
-          <TableCaption>
-            <TableCaption.Title>No images</TableCaption.Title>
-            <TableCaption.Description>
-              There are no images stored in Site Manager at the moment. You can
-              either upload images, or connect to an upstream image source to
-              download images from.
-            </TableCaption.Description>
-          </TableCaption>
-        }
-        pin={[
-          { value: "Ubuntu", isTop: true },
-          { value: "Other", isTop: false },
-        ]}
-        rowSelection={selectedRows}
-        setRowSelection={setSelectedRows}
-        sortBy={[{ id: "release", desc: true }]}
-      />
-    </>
+    <GenericTable
+      canSelect
+      columns={columns}
+      data={images}
+      filterCells={filterCells}
+      filterHeaders={filterHeaders}
+      groupBy={["name"]}
+      noData={
+        <TableCaption>
+          <TableCaption.Title>No images</TableCaption.Title>
+          <TableCaption.Description>
+            There are no images stored in Site Manager at the moment. You can
+            either upload images, or connect to an upstream image source to
+            download images from.
+          </TableCaption.Description>
+        </TableCaption>
+      }
+      pin={[
+        { value: "Ubuntu", isTop: true },
+        { value: "Other", isTop: false },
+      ]}
+      rowSelection={selectedRows}
+      setRowSelection={setSelectedRows}
+      sortBy={[{ id: "release", desc: true }]}
+    />
   );
 };
