+++ src/app/images/views/ImageList/ImageListHeader/_index.scss
@@ -1,0 +1,3 @@
+.stop-image-import {
+  margin-left: 1rem;
+}
